+++ Sources/ModelsR5/Base64Binary.swift
@@ -24,7 +24,7 @@ import FMCore
  
  http://hl7.org/fhir/datatypes.html#base64Binary
  */
-public struct Base64Binary: FHIRPrimitiveType {
+public struct Base64Binary: FHIRPrimitiveType, Sendable {
 	
 	public var dataString: String